requirements.txt
streamlit==1.41.1
flask==3.1.0
gunicorn==23.0.0
requests==2.32.3
plotly==5.24.1
pandas==2.2.3
pytz==2024.2
anthropic>=0.42.0
google-cloud-storage>=2.14.0
google-ads>=29.0.0
